- Fix (hopefully) all warnings during compilation - Basic support for LPRng printers installed using YaST   (only useful for printer listing)
svn path=/trunk/kdelibs/; revision=139718

+++ kdeprint/cups/cupsaddsmb2.cpp
@@ -196,7 +196,7 @@ void CupsAddSmb::checkActionStatus()
 
 void CupsAddSmb::nextAction()
 {
-	if (m_actionindex < m_actions.count())
+	if (m_actionindex < (int)(m_actions.count()))
 		QTimer::singleShot(1, this, SLOT(doNextAction()));
 }
